+++ rmk-macro/src/default_config/mod.rs
@@ -1,4 +1,5 @@
 pub(crate) mod esp32;
+pub(crate) mod nrf52810;
 pub(crate) mod nrf52832;
 pub(crate) mod nrf52840;
 pub(crate) mod rp2040;